+++ CYYVG6_HFT_2021221.Logic/StudentLogic.cs
@@ -28,6 +28,10 @@ namespace CYYVG6_HFT_2021221.Logic
 
         public void Delete(int id)
         {
+            if (id < 0)
+            {
+                throw new ArgumentException("Invalid Student ID");
+            }
             studentRepository.Delete(id);
         }
